[#59] refactor : 필요 없는 파일 및 주석 삭제
